+++ common/src/main/java/com/gitlab/cdagaming/craftpresence/utils/gui/impl/SelectorGui.java
@@ -205,7 +205,7 @@ public class SelectorGui extends ExtendedScreen {
                     new ScrollableListControl(
                             mc, this,
                             getScreenWidth(), getScreenHeight(),
-                            32, getScreenHeight() - 45, renderType != RenderType.None && !CraftPresence.CONFIG.accessibilitySettings.stripExtraGuiElements ? 45 : 18,
+                            32, getScreenHeight() - 45,
                             itemList, originalValue,
                             renderType
                     )